Reduce Zookeeper prompt test lint pressure
